add QFT to builders
