+++ src/tensor_equiv/__init__.py
@@ -3,6 +3,7 @@ from .builders import (
     get_n_bell_pairs_circuit,
     get_choi_state_circuit,
     get_diagonal_choi_state_circuit,
+    build_qft_circuit,
 )
 
 
@@ -22,4 +23,5 @@ __all__ = [
     "check_equivalence",
     "check_equivalence_with_ancillas",
     "REPLACE_CONDITIONALS",
+    "build_qft_circuit",
 ]
